requirements.txt
absl-py==2.2.2
ale-py==0.10.2
appnope==0.1.4
asttokens==3.0.0
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
rich==14.0.0
setuptools==78.1.0
Shimmy==2.0.0
six==1.17.0
stable_baselines3==2.4.1
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
PyYAML==6.0.2
